requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

streamlit==1.32.2

torch==2.2.2
transformers==4.38.2
sentencepiece==0.1.99

chromadb==0.4.24
sentence-transformers==2.6.1

nltk==3.8.1
tiktoken==0.6.0
pandas==2.2.1
pyyaml==6.0.1
requests==2.31.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

pydantic==2.6.4
